[FLINK-5522] [storm compatibility] Move Storm LocalCluster based test to a separate class
This fixes the problem that the Storm LocalCluster can't run with powermock

This closes #3138
